Decouples ShowcaseCard component
