+++ src/components/showcase/ShowcaseCard/components/Button.js
@@ -1,0 +1,14 @@
+import React from 'react'
+
+const Button = ({text,hrefValue}) => (
+  <a
+  className="button button--small button--secondary button--block"
+  href={hrefValue}
+  target="_blank"
+  rel="noreferrer noopener"
+>
+  {text}
+</a>
+)
+
+export default Button
